chores: little by little
