+++ backend/updates/routing.py
@@ -1,11 +1,11 @@
 from django.urls import re_path
 
-from .consumers import RealtimeConsumer
+from .consumers import UpdatesConsumer
 
 websocket_urlpatterns = [
     re_path(
         r"ws/updates/$",
-        RealtimeConsumer.as_asgi(),
+        UpdatesConsumer.as_asgi(),
     ),
 ]
